Move boilerplate to an abstract class
